requirements/requirements.txt
chatterbot==1.0.8
chatterbot_corpus==1.2.0
flask==2.2.5
spacy==2.0.16
requests
PyGithub
google-api-python-client
nltk
beautifulsoup4
SQLAlchemy
pymongo

